[BE-S14P31A204-306] fix: progress type과 worker schema 정렬

+++ BE/src/main/java/com/a204/batang/domain/render/messaging/dto/SdRenderEventMessage.java
@@ -38,7 +38,7 @@ public record SdRenderEventMessage(
         @JsonProperty("output_artifact_id")
         UUID outputArtifactId,
         String status,
-        Integer progress,
+        Double progress,
         Map<String, Object> output,
         SdRenderError error,
         @JsonProperty("idempotency_key")
